fastapi-sports-scores/requirements.txt
fastapi
uvicorn
httpx
aiohttp
python-multipart
cachetools
redis
requests
Pillow